Change `parse_alg!(…)` and `parse_move!(…)` to return a lazily constructed shared reference.
This avoids duplicate parsing and duplicated memory for the same alg/move.

This also makes it more convenient to pass to methods that take a reference, as you don't have to add `&`.

For cases where you still need to own a copy, you can still create a copy (without re-parsing every time).